Updated the code based on some debugging done today. The task so far has just been for inputs and seems to work. The timing for the sample task and send task needs to be fixed, I just wanted to check this in since it works.
Updated all of the messages in message class to use the header verification vs the simple byte verification. Found a bug in that code that caused all messages to be invalid. The compare was being done with the first 8 bytes in the class vice the first 8 bytes of the message. That is fixed now!

Updated the serial delay computation to be correct. 

All of the sizeof functions have been changed to use the type vice the data element... I am not sure why I used the data element since that is never a good idea.

Added the gedit temp files to the gitignore file.